+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - PR #55 - Update notebooks to use timeit instead of time
 - PR #56 - Ability to precompile select Numba/CuPy kernel on import
 - PR #60 - Updated decimate function to use an existing FIR window
+- PR #65 - Added deprecation warning for Numba kernels
 
 ## Bug Fixes
 - PR #44 - Fix issues in pytests